Streaming sites be like

+++ database/migrations/2016_07_05_223706_create_videos_table.php
@@ -20,7 +20,7 @@ class CreateVideosTable extends Migration
             $table->enum('translation_type', ['sub', 'dub']);           // episode
             $table->integer('mirror')->unsinged();                      // video
 
-            $table->integer('episode_num');                             // episode
+            $table->float('episode_num');                               // episode
             $table->string('link_stream', 1024);                        // video
             $table->string('link_episode', 1024);                       // episode
             $table->string('notes');                                    // video
